pyproject.toml
[tool.poetry]
name = "phantom-sc2"
version = "2.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "ares-sc2/src/ares"}]

[tool.poetry.dependencies]
ares-sc2 = { path = "ares-sc2", develop = false }
burnysc2 = { git = "https://github.com/august-k/python-sc2", branch = "develop" }
python = ">=3.11, <3.12"
black = "^24.4.2"
isort = "^5.12.0"
requests = "^2.32.3"
scikit-image = "^0.24.0"
sc2-helper = "^0.2.1"
jupyter = "^1.0.0"
snakeviz = "^2.2.0"
cython = "^3.0.10"
line-profiler = "^4.1.3"
setuptools = "^75.3.0"
flake8 = "^7.1.0"
flake8-pyproject = "^1.2.3"
flake8-isort = "^6.1.1"
mypy = "^1.12.0"
autoflake = "^2.3.1"
mpyq = "^0.2.5"
s2protocol = "^5.0.13.92440.4"
pandas = "^2.2.3"
tqdm = "^4.67.0"
types-tqdm = "^4.66.0.20240417"


[build-system]
requires = ["poetry-core", "Cython", "numpy", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.autoflake]
remove-all-unused-imports = true

[tool.flake8]
max-line-length = 120
ignore = ["E203", "W503"]
exclude = [
    "ares-sc2",
]

[tool.black]
extend-exclude = 'ares-sc2/*'
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"
skip_glob = ["ares-sc2/*"]

[[tool.mypy.overrides]]
module = [
    "ares.*",
    "sc2.*",
    "scipy.*",
    "pandas.*",
    "sklearn.*",
    "cython_extensions.*",
    "mpyq.*",
    "s2protocol.*",
]
ignore_missing_imports = true